0.6.12-alt1.2
- Update to 6.12.7
- fix threads
- add configure

+++ qt4-fsarchiver/src/regmulti.c
@@ -42,7 +42,7 @@ int regmulti_empty(cregmulti *m)
     m->count=0;
     m->usedsize=0;
     
-    for (i=0; i < m->maxitems; i++)
+    for (i=0; (unsigned int) i < m->maxitems; i++)
         m->objhead[i]=NULL;
     
     return 0;
